Make testsuite flake8 compliant
Involves a bit of refactor to tests/util/__init__.py to break it down.
