+++ tests/system/dir_test.py
@@ -1,7 +1,4 @@
-#!/usr/bin/python
-
 import os
-import mock
 from nose.tools import istest
 
 from tests import system
