+++ tests/unit/action/backup_test.py
@@ -5,7 +5,7 @@ from nose.tools import istest
 from salve import action
 from salve.action import backup
 from salve.filesys import ConcreteFilesys
-from salve.context import ExecutionContext, FileContext
+from salve.context import ExecutionContext
 
 from tests.util import ensure_except, scratch, full_path
 from tests.unit.action import dummy_file_context
